Move the performance tests to their own file and predicate them on an environment variable secondary to the "is debug build" check. Skip them using XCTSkipUnless() instead of manually returning.
